backend/pyproject.toml
[project]
name = "launchboard-backend"
version = "0.2.0"
description = "Launchboard — AI job search agent backend"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.109.0",
    "PyJWT[crypto]>=2.10.0",
    "psycopg[binary]>=3.2.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0",
    "python-jobspy>=1.1.75",
    "PyPDF2>=3.0.0",
    "openai>=1.0.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
